front: split NGE-to-OSRD note-related logic in separate file
